ARMv7: start.S: stay in HYP mode if u-boot is entered in it
The KVM and Xen hypervisors for the Cortex-A15 virtualization
implementation need to be entered in HYP mode. Should the primary
board firmware already enter HYP mode (Calxeda firmware does that),
we should not deliberately drop back to SVC mode.
Since U-boot does not use the MMU, running in HYP mode is just fine.

Change-Id: Ibb52de3484e51ba0f81e7f0029cec2de5c4da911
Signed-off-by: Andre Przywara <andre.przywara@linaro.org>