vboot: image: Add support for signing of FIT configurations
While signing images is useful, it does not provide complete protection
against several types of attack. For example, it it possible to create a
FIT with the same signed images, but with the configuration changed such
that a different one is selected (mix and match attack). It is also possible
to substitute a signed image from an older FIT version into a newer FIT
(roll-back attack).

Add support for signing of FIT configurations using the libfdt's region
support.

Please see doc/uImage.FIT/signature.txt for more information.

Change-Id: I0b81c3cc3b50c66ea74efd3a2385a1c8735a4d0b
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Jaehoon Kim <jh228.kim@samsung.com>