vboot: Add verified boot information and test
Add a description of how to implement verified boot using signed FIT images,
and a simple test which verifies operation on sandbox.

The test signs a FIT image and verifies it, then signs a FIT configuration
and verifies it. Then it corrupts the signature to check that this is
detected.

Change-Id: If45c44ca927853118c7ad3033b3b0c80a9fa331b
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Jaehoon Kim <jh228.kim@samsung.com>