artik: configs: fix boot partition offset
This patch fixes boot partition offset of DFU for the Thor downloader.
CONFIG_DFU_ALT_BOOT_SD information was added to support booting by SD
card. But, there was an error because this feature was under developing at
that time.
EMMC and SD have differnt offset because of the difference between GPT
and MBR. EMMC and SD use GPT and MBR, respectively. GPT have a partition
for boot separately. So, EMMC don't need offset. But, SD need one block as
a offset for MBR information.

Change-Id: I0c16c4fc8466d6c22567b43fe357c3354c0b0381
Signed-off-by: jino.cho <jino.cho@samsung.com>