vboot: fdt: Check for a token to skip auto-hash validation
Allow the itb file to declare to u-boot that its hash should not be
checked automatically on bootm or iminfo.  This allows an image to
either be checked automatically or to include a script which may
check it otherwise (such as after part of the itb has been relocated
to RAM by the script).

Change-Id: I0e8feb6ad18d3339595954556422d797c39b60a3
Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>
Signed-off-by: Jaehoon Kim <jh228.kim@samsung.com>