vboot: image: Introduce fit_image_load() to load images from FITs
At present code to load an image from a FIT is duplicated in the three
places where it is needed (kernel, fdt, ramdisk).

The differences between these different code copies is fairly minor.
Create a new function in the fit code which can handle any of the
requirements of those cases.

Change-Id: I3d67afa2c7b3f6651a2c3b1297f25455db0e1cdb
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Jaehoon Kim <jh228.kim@samsung.com>