artik053: enable pull-up on gpg16
Unlike ARTIK-051 variants, GPG16 pin of ARTIK-053 is pulled-up by the
external circuit and its value has logical high when the button attached
to it is not pressed. This is problematic when we are using the U-BOOT
binary for ARTIK-053 on ARTIK-051, because GPG16 on ARTIK-051 board is
pulled-down by default and thus U-BOOT recognizes the button is pressed
down.

To workaround this, we need to make GPG16 be pulled-up temporarily while
we are checking the GPG16, which is used as a button for triggering the
factory reset, though it is a little bit dirty.

Change-Id: Ieaabdde4d2ea17daf22bc947fa38bd53bd86dcad
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>