arm: nexell: do not use global variables before relocation
We should avoid using any global variables before the relocation where
the .bss section is cleared. Those still need to be kept before the
initialization of .bss could use the "global_data" structure which is
available before and after the relocation. Timestamp and lastdec are
also appropriate to do so.

This commit fixes the FDT corruption issue when CONFIG_OF_SEPARATE=y.

Change-Id: I9e103af81e73aa5c3f2b8acb6a0ea23acd7f798e
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>