artik: configs: remove CONFIG_FIT_BEST_MATCH
As of now, on all ARTIK platforms, booting from FIT is not supported,
though CONFIG_FIT is set. I do not know historical backgrounds of why
it is currently set, but actually it has no effect. Therefore, having
CONFIG_FIT_BEST_MATCH=y set also makes no sense in that we do not boot
from FIT as of now. We have a plan to use CONFIG_FIT in the near future,
but CONFIG_FIT_BEST_MATCH will never be used, as it prevents us from
being fallback to default configuration node in FIT. It seems a dumb
copy and paste mistake while copying from other boards. So, let's remove
it.

Change-Id: I907b129325fba0f9737acdfbf33f2c3892ad9786
Signed-off-by: Heesub Shin <heesub.shin@samsung.com>