mmc: nexell_dw_mmc: convert to Driver Model driver
This patch converts the old-style nexell mmc driver to DM(Driver model)
based driver. Unfortunately, it does not mean it can support CONFIG_BLK
because this version of u-boot does not support the mmc block to
support it. By converting it to DM-style, we can integrate the pinctrl
driver with this driver. It makes to control pin mux/control more
easier than old-style.

Change-Id: I9ec5bdc50c082aa00def2dc34114100aaf90286a
Signed-off-by: Chanho Park <chanho61.park@samsung.com>