board: nexell: artik711s: choose sdcard/clipper1 by bootmode
The sdcard of artik711s raptor board is muxed with the ccir1 camera.
It can be switched by jumper select but there is no way to control the
jumper by S/W. Thus, we decided to switch them according to boot mode.
When a board is turned on eMMC boot mode,
 - sdcard: disable
 - clipper1: enable
otherwise,
 - sdcard: enable
 - clipper1: disable

Change-Id: I0ef9674b8865d943394a27a8aba5c80fe30994ed
Signed-off-by: Chanho Park <chanho61.park@samsung.com>