arm: dts: artik711s: modify shift of eMMC
This patch modifies drive_shift and sample_shift of eMMC.

The drive_shift can have values as 0 to 3. The value "0" means
degree 0, "1" means degree 90, "2" means degree 180 and "3" means
degree 270. In the same as drive_shift, the sample_shift can have
values as 0 to 3. The meanings of values are also the same.

There is sometimes eMMC read fail during booting, and it occured
because of end-bit error. This patch solves it.

Change-Id: I3bd771db851e33da64c233a9ff369e2226deabec
Signed-off-by: Sowon Na <sowon.na@samsung.com>