arm: dts: artik710: modify shift of eMMC
This patch modifies drive_shift and sample_shift of eMMC.

The drive_shift can have values as 0 to 3. The value "0" means
degree 0, "1" means degree 90, "2" means degree 180 and "3" means
degree 270. In the same as drive_shift, the sample_shift can have
values as 0 to 3. The meanings of values are also the same.

There is sometimes eMMC read fail during booting, and it occured
because of end-bit error. This patch solves it.

Change-Id: I06e6a840df4ecac7d5c82e40b74ebc580b9027c0
Signed-off-by: Sowon Na <sowon.na@samsung.com>