arm: dts: artik530: optimize shift of eMMC
This patch modifies drive_shift and sample_shift of eMMC.
The drive_shift can have from 0 to 3 values. The meanings of the
values are degree of the phase shifted clock.

0: 0 degree phase shifted clock
1: 90 degree phase shifted clock
2: 180 degree phase shifted clock
3: 270 degree phase shifted clock

In the same as drive_shift, the sample_shift can have values as 0
to 3. The meanings of values are also the same.

There are sometimes fastboot fail with some targets, and this patch
solves it.

Change-Id: Id375b393e045a1dde41c9321900571b37dc76953
Signed-off-by: Sowon Na <sowon.na@samsung.com>